requirements.txt
afdko==4.0.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
booleanOperations==0.9.0
brotli==1.2.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
defcon==0.12.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
fontMath==0.9.4
fontParts==0.13.3
fontPens==0.2.4
fonttools==4.61.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
MutatorMath==3.0.1
pyclipper==1.4.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.55.0
shellingham==1.5.4
starlette==1.0.0
tqdm==4.67.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ufonormalizer==0.6.3
ufoProcessor==1.14.1
unicodedata2==17.0.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
zopfli==0.4.0
google-genai==1.13.0
